% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Sample Medical Dataset for TestTrek}
\format{
A data frame with 500 rows and the following columns:
\describe{
  \item{condition}{Medical condition of the patient (nominal: Yes/No).}
  \item{blood_type}{Blood type of the patient (nominal: A, B, AB, O).}
  \item{pain_scale}{Pain scale rated by the patient (ordinal: 1-10 scale).}
  \item{hospital_visits}{Number of hospital visits in the past year (discrete).}
  \item{age}{Age of the patient (continuous).}
  \item{weight}{Weight of the patient in kilograms (continuous).}
  \item{bp_baseline}{Baseline blood pressure measurement (continuous).}
  \item{bp_6months}{Blood pressure measurement after 6 months from the baseline (continuous).}
  \item{bp_12months}{Blood pressure measurement after 12 months from the baseline (continuous).}
}
}
\source{
Source: [none]
}
\usage{
data
}
\description{
A dataset containing various medical variables for statistical test demonstration.
}
\examples{
head(data)
}
\references{
None
}
\keyword{datasets}
