% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recommend_test_dependent}
\alias{recommend_test_dependent}
\title{Recommend Appropriate Statistical Test for Dependent Samples}
\usage{
recommend_test_dependent(data, var1, var2)
}
\arguments{
\item{data}{A data frame containing the variables to analyze.}

\item{var1}{The name of the first variable.}

\item{var2}{The name of the second variable.}
}
\value{
Prints the recommended statistical test and effect size.
}
\description{
This function analyses two variables from a dataset and recommends
the appropriate statistical test and effect size for dependent samples.
}
\examples{
\dontrun{
recommend_test_dependent(my_data, "var1", "var2")
}

}
