% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recommend_test_independent}
\alias{recommend_test_independent}
\title{Recommend Appropriate Statistical Test for Independent Samples}
\usage{
recommend_test_independent(data, var1, var2)
}
\arguments{
\item{data}{A data frame containing the variables to analyze.}

\item{var1}{The name of the first variable.}

\item{var2}{The name of the second variable.}
}
\value{
Prints the recommended statistical test, effect size,
and post-hoc tests (when necessary).
}
\description{
This function analyses two variables from a dataset and recommends
the appropriate statistical test, effect size, and post-hoc tests
(when necessary) for independent samples.
}
\examples{
\dontrun{
recommend_test_independent(my_data, "var1", "var2")
}

}
